Fixed validation on all Edit forms, changed prices to use decimals

+++ CoreTemplate/CoreTemplate.Accessors/Models/DTO/MovieDTO.cs
@@ -14,9 +14,9 @@ namespace CoreTemplate.Accessors.Models.DTO
 
         public string ImageUrl { get; set; }
 
-        public int PurchaseCost { get; set; }
+        public decimal PurchaseCost { get; set; }
 
-        public int RentCost { get; set; }
+        public decimal RentCost { get; set; }
 
         public ICollection<MovieGenreDTO> Genres { get; set; }
